requirements.txt
aiogram==2.25.1
openpyxl==3.1.5
python-dotenv==1.0.1
